Dynamic-access coverage for jakarta.servlet.jsp.jstl.fmt.LocaleSupport (2/2)
